requirements.txt
reportlab
python-docx
google-generativeai
pydantic
python-dotenv
